pyproject.toml
[project]
name = "timbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "numpy>=1.24.0",
    "piper-tts>=1.3.0",
    "ruff>=0.12.12",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "tqdm>=4.67.1",
    "typer>=0.17.4",
    "faster-whisper",
    "opencv-python>=4.8.0",
    "ollama>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/timbot"]

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "ty>=0.0.1a20",
]
